address comments - extract @Composables in separate files, fix icons issue, use repeat, remove comments
